+++ docs/changelog.rst
@@ -16,6 +16,9 @@ Added
 
 Changed
 -------
+- New window smoothing algorithm in `setup_rivers` to avoid cells with small river length. 
+  Set the min_rivlen_ratio argument to a value larger than zero to apply the smoothing.
+  Note: requires pyflwdir v0.5.6 `PR #92 <https://github.com/Deltares/hydromt_wflow/pull/92>`_
 
 Fixed
 -----